api/terminal/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
talos-api
dogpile.cache
gunicorn
paramiko
tornado
pyte
requests
pymysql
pyjwt<2.0.0
pycrypto
gevent
pytz
apscheduler